Removes test files from production builds
