+++ apps/judicial-system/web-e2e/src/integration/Court/InvestigationCases/receptionAndAssignment.spec.ts
@@ -1,14 +1,10 @@
-import {
-  makeCourt,
-  makeCustodyCase,
-} from '@island.is/judicial-system/formatters'
 import { CaseState } from '@island.is/judicial-system/types'
 import {
   IC_OVERVIEW_ROUTE,
   IC_RECEPTION_AND_ASSIGNMENT_ROUTE,
 } from '@island.is/judicial-system/consts'
 
-import { intercept } from '../../../utils'
+import { makeCourt, makeCustodyCase, intercept } from '../../../utils'
 
 describe(`${IC_RECEPTION_AND_ASSIGNMENT_ROUTE}/:id`, () => {
   beforeEach(() => {
