+++ apps/judicial-system/web-e2e/src/integration/Court/RestrictionCases/receptionAndAssignment.spec.ts
@@ -1,14 +1,10 @@
-import {
-  makeCourt,
-  makeCustodyCase,
-} from '@island.is/judicial-system/formatters'
 import { CaseState } from '@island.is/judicial-system/types'
 import {
   OVERVIEW_ROUTE,
   RECEPTION_AND_ASSIGNMENT_ROUTE,
 } from '@island.is/judicial-system/consts'
 
-import { intercept } from '../../../utils'
+import { makeCustodyCase, makeCourt, intercept } from '../../../utils'
 
 describe(`${RECEPTION_AND_ASSIGNMENT_ROUTE}/:id`, () => {
   beforeEach(() => {
