+++ apps/judicial-system/web/tsconfig.app.json
@@ -4,6 +4,6 @@
     "outDir": "../../../dist/out-tsc",
     "types": ["node", "jest"]
   },
-  "exclude": ["**/*.spec.ts", "**/*.spec.tsx"],
+  "exclude": ["**/*.spec.ts", "**/*.spec.tsx", "**/mocks.ts"],
   "include": ["**/*.ts", "**/*.tsx", "next-env.d.ts"]
 }
