+++ libs/judicial-system/formatters/src/index.ts
@@ -1,2 +1,1 @@
 export * from './lib/formatters'
-export * from './lib/testData'
